LNC 2025/Kid Shop/src/requirements.txt
flask
pyjwt